+++ template_variables/RELEASES.md
@@ -203,3 +203,11 @@ Based on:
 - Speakeasy CLI 1.189.0 (2.263.3) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v3.0.3] template_variables
+
+## 2024-02-24 00:20:09
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/template-variables.yaml
+- Speakeasy CLI 1.193.4 (2.272.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.4] template_variables